git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/branches/KinodynamicRRT@962 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/MotionPlanning/AvoidRegionValidity.h
@@ -73,7 +73,7 @@ IsValidImpl(CfgType& _cfg, CDInfo& _cdInfo, const string& _callName) {
 
     //if center is outside boundary, check each component of the robot to ensure
     //that none lie within the avoid region
-    for(int m = 0; m < robot->NumFreeBody(); ++m) {
+    for(size_t m = 0; m < robot->NumFreeBody(); ++m) {
       typedef vector<Vector3d>::const_iterator VIT;
       Transformation& worldTransformation = robot->GetFreeBody(m)->
         WorldTransformation();
